requirements.txt
paho-mqtt==1.6.1
google-cloud-pubsub==2.13.0
Flask==2.0.1